take care of doulbe-to-float conversion warnings; add in some static casts for SnapInt

+++ neo/idlib/math/Vector.h
@@ -218,8 +218,8 @@ ID_INLINE void idVec2::Clamp( const idVec2 &min, const idVec2 &max ) {
 }
 
 ID_INLINE void idVec2::Snap() {
-	x = floor( x + 0.5f );
-	y = floor( y + 0.5f );
+	x = floorf( x + 0.5f );
+	y = floorf( y + 0.5f );
 }
 
 ID_INLINE void idVec2::SnapInt() {
@@ -693,15 +693,15 @@ ID_INLINE void idVec3::Clamp( const idVec3 &min, const idVec3 &max ) {
 }
 
 ID_INLINE void idVec3::Snap() {
-	x = floor( x + 0.5f );
-	y = floor( y + 0.5f );
-	z = floor( z + 0.5f );
+	x = floorf( x + 0.5f );
+	y = floorf( y + 0.5f );
+	z = floorf( z + 0.5f );
 }
 
 ID_INLINE void idVec3::SnapInt() {
-	x = float( int( x ) );
-	y = float( int( y ) );
-	z = float( int( z ) );
+	x = static_cast<float>( static_cast<int>( x ) );
+	y = static_cast<float>( static_cast<int>( y ) );
+	z = static_cast<float>( static_cast<int>( z ) );
 }
 
 ID_INLINE int idVec3::GetDimension() const {
